+++ projects/star/train_atom.py
@@ -24,7 +24,7 @@ from mmengine import mkdir_or_exist
 # pose estimation
 from cryostar.utils.transforms import SpatialGridTranslate
 # other
-from cryostar.dataio import StarfileDataSet, StarfileDatasetConfig, Mask
+from cryostar.utils.dataio import StarfileDataSet, StarfileDatasetConfig, Mask
 from cryostar.utils.ctf_utils import CTFRelion
 from cryostar.utils.misc import log_to_current, \
     pl_init_exp, pretty_dict, set_seed, calc_cor_loss, calc_kl_loss
